requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
hypothesis>=6.82.0
pytest>=7.4.0
pytest-cov>=4.1.0